# improve performance (+99.5% throughput)