@AlCalzone 2023-03-16
### Second change: Use a string search instead of RegExp to match ansi codes
Result:
```
  original:
    23 692 ops/s, ±0.73%   | slowest, 17.61% slower

  for loop:
    27 710 ops/s, ±0.94%   | 3.64% slower

  for-loop + no regex:
    28 756 ops/s, ±0.58%   | fastest
```