@AlCalzone 2023-03-16
### 4th change: Avoid converting ansi codes to number for end code lookup
Result:
```
  original:
    27 402 ops/s, ±0.76%   | slowest, 41.57% slower
    
  (snip)

  no regex + no spread:
    43 911 ops/s, ±0.81%   | 6.36% slower

  no conversion between number and string:
    46 895 ops/s, ±0.98%   | fastest
```