@AlCalzone 2023-03-16
### 5th change: Don't use string.split to access first char before first ";", use indexOf and slice
Result:
```
  original:
    22 828 ops/s, ±1.84%   | slowest, 49.89% slower

  (snip)

  no conversion between number and string:
    42 327 ops/s, ±0.90%   | 7.08% slower

  indexOf instead of splitting:
    45 553 ops/s, ±0.90%   | fastest
```