@AlCalzone 2023-03-18
### Last change: "Tokenize" input into ansi codes and characters, operate on pre-analyzed array
This one is much more involved. I basically rewrote everything to first analyze the input string and turn it into an array of ANSI codes and characters. Slicing is then done by operating on the array.

I had to change two tests though - one unnecessarily used two separate foreground colors where the 2nd overwrote the 1st. And in another one, the expected string had the end codes in the same order as the start codes, while all others had it in the opposite order. I didn't see any difference when printing those strings.

This change has a lot of potential though:
```
  original:
    20 822 ops/s, ±1.70%
  
  (snip)

  indexOf instead of splitting:
    41 278 ops/s, ±4.02%

  tokenize:
    51 791 ops/s, ±2.93%
```

and if the result of the tokenization is memorized (export `tokenize` function, and add a copy of the `sliceAnsi` function which operates on a token array, instead of a string) , repeated slices on the same input (like in the test case) are much faster:

```
  tokenize, the smart way:
    84 850 ops/s, ±3.21%
```